Revert "Fix for #211"
This reverts commit 62998924a840cae3f7ff193e514eda8589dc6df2.

Breaks other things (#212)

+++ build.sh
@@ -184,12 +184,14 @@ fi
 
 pushd $SDK_DIR/MacOSX$SDK_VERSION.sdk &>/dev/null
 set +e
-files=$(echo $SDK_DIR/MacOSX$SDK_VERSION.sdk/System/Library/Frameworks/Kernel.framework/Versions/A/Headers/std*.h)
-for file in $files; do
-  rm -f usr/include/$(basename $file)
-  cp $file usr/include
-done
-cp -f $BASE_DIR/oclang/quirks/tgmath.h usr/include
+if [ $PLATFORM == "FreeBSD" ] || [ $PLATFORM == "OpenBSD" ]; then
+  files=$(echo $SDK_DIR/MacOSX$SDK_VERSION.sdk/System/Library/Frameworks/Kernel.framework/Versions/A/Headers/std*.h)
+  for file in $files; do
+    rm -f usr/include/$(basename $file)
+    cp $file usr/include
+  done
+  cp -f $BASE_DIR/oclang/quirks/tgmath.h usr/include
+fi
 if [ ! -f "usr/include/float.h" ]; then
   cp -f $BASE_DIR/oclang/quirks/float.h usr/include
 fi
